+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Components
+
+- Combobox
+- TimePicker with combobox
+- Date picker
+- Date range picker
+- File input
+- Character count
+
+## [1.1.0](https://github.com/metonym/svelte-uswds/releases/tag/v1.1.0) - 2020-09-27
+
+**Features**
+
+- Add `TimePicker`, `StepIndicator`, `Identifier` components
+
 ## [1.0.0](https://github.com/metonym/svelte-uswds/releases/tag/v1.0.0) - 2020-09-20
 
 **Features**